Improved Plantard Arithmetic for Lattice-based Cryptography (#244)
* integrate improved Plantard arithmetic into Kyber

* support m4fspeed and rename m4plantard to m4fstack

* remove obj file

* add benchmarks

+++ crypto_kem/kyber768/m4fspeed/ntt.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-extern const int16_t zetas[64];
+extern const int32_t zetas[64];
 
 void ntt(int16_t *poly);
 void invntt(int16_t *poly);
